feat(api): publish report and error contracts
