+++ docs/user-guide/user-guide.html
@@ -109,7 +109,7 @@ func handler(w http.ResponseWriter, r *http.Request) {
   "entry_call": { "function_name": "issueToken" },
   "supporting_calls": []
 }</code></pre></div></div></div><p class="reachability-note">These examples are synthetic and abbreviated. Exact field names and nesting depend on the artifact. <code>finding_id</code> joins the records; <code>source</code>, file, line, match, and <code>metadata</code> describe the evidence. Dependency-backed records may also include <code>dependency_info</code>. Reachability exports can add <code>call_chains</code>, <code>entry_call</code>, <code>crypto_call</code>, and <code>supporting_calls</code>. If a link is incomplete, <code>unresolved_reason</code> and related caveats explain why. Treat every field as review evidence, not a runtime proof or policy verdict.</p><div class="field-grid"><div class="field"><code>finding_id</code><p>Join key shared with the interim finding.</p></div><div class="field"><code>call_chains</code><p>Static path or paths from a reachable caller to the finding anchor.</p></div><div class="field"><code>entry_call</code><p>How execution entered the next function, including best-effort parameter provenance.</p></div><div class="field"><code>crypto_call</code><p>The matched crypto-relevant call at the end of a chain.</p></div><div class="field"><code>supporting_calls</code><p>Context and lifecycle calls derived from the same graph. They are not findings.</p></div><div class="field"><code>unresolved_reason</code><p>Why a finding could not be fully connected. Investigate rather than treating it as unreachable.</p></div></div></section>
-<section class="section" id="outputs" data-search="outputs interim JSON CycloneDX CBOM graph fragment format schema 1.4 1.6"><div class="section-head"><div class="eyebrow">Evidence formats</div><h2>Pick the artifact for the reader.</h2></div><p class="lede">Keep interim JSON for detailed analysis. Use CycloneDX CBOM for standardized reporting. Add callgraph or graph fragment exports when structural context or reuse matters.</p><div class="three-col"><div class="mini-panel"><h3>Interim JSON</h3><p>Detailed findings, source match, metadata, status, attribution, and <code>finding_id</code>. Format version is <code>1.4</code>.</p></div><div class="mini-panel"><h3>CycloneDX CBOM</h3><p>Standardized Cryptography Bill of Materials. The documented output is CycloneDX <code>1.6</code>.</p></div><div class="mini-panel"><h3>Structure exports</h3><p>Finding-centric callgraph export uses schema <code>6.7</code>. Findings carry <code>reachable</code> when the scan included dependencies: <code>true</code> when user code reaches the crypto, <code>false</code> when it does not. The field is absent when a library is scanned on its own, since there is no user code to reach from. Reusable fragments use <code>graph-fragment-1.8</code>.</p></div></div><div class="two-col"><div><h3>Interim JSON</h3><div class="code-card"><div class="code-head"><span>findings.json</span><button class="copy" type="button">Copy</button></div><pre><code>{
+<section class="section" id="outputs" data-search="outputs interim JSON CycloneDX CBOM graph fragment format schema 1.4 1.6"><div class="section-head"><div class="eyebrow">Evidence formats</div><h2>Pick the artifact for the reader.</h2></div><p class="lede">Keep interim JSON for detailed analysis. Use CycloneDX CBOM for standardized reporting. Add callgraph or graph fragment exports when structural context or reuse matters.</p><p>Go integrations can import <code>pkg/schema</code> for the interim report and <code>pkg/failure</code> for structured error payloads. Both preserve the documented JSON contract without requiring implementation-package imports.</p><div class="three-col"><div class="mini-panel"><h3>Interim JSON</h3><p>Detailed findings, source match, metadata, status, attribution, and <code>finding_id</code>. Format version is <code>1.4</code>.</p></div><div class="mini-panel"><h3>CycloneDX CBOM</h3><p>Standardized Cryptography Bill of Materials. The documented output is CycloneDX <code>1.6</code>.</p></div><div class="mini-panel"><h3>Structure exports</h3><p>Finding-centric callgraph export uses schema <code>6.7</code>. Findings carry <code>reachable</code> when the scan included dependencies: <code>true</code> when user code reaches the crypto, <code>false</code> when it does not. The field is absent when a library is scanned on its own, since there is no user code to reach from. Reusable fragments use <code>graph-fragment-1.8</code>.</p></div></div><div class="two-col"><div><h3>Interim JSON</h3><div class="code-card"><div class="code-head"><span>findings.json</span><button class="copy" type="button">Copy</button></div><pre><code>{
   "version": "1.4",
   "findings": [{
     "file_path": "src/crypto/cipher.go",
